feat: (almost) completes browse restaurants (Page H)
